requirements.txt
asgiref==3.8.1 
django==5.1.2 
sqlparse==0.5.1 
typing-extensions==4.12.2
dj-database-url==2.2.0
django-environ==0.11.2
python-dotenv==1.0.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
PyJWT==2.9.0 
PyYAML==6.0.2 
ShopifyAPI==12.6.0 
pyactiveresource==2.2.2 
six==1.16.0
python-decouple==3.8
certifi==2024.8.30 
charset-normalizer==3.4.0 
idna==3.10 
requests==2.32.3 
urllib3==2.2.3
pinecone-client==5.0.1 
pinecone-plugin-inference==1.1.0 
pinecone-plugin-interface==0.0.7 
tqdm==4.66.5
filelock==3.16.1 
fsspec==2024.9.0 
huggingface-hub==0.25.2 
numpy==2.1.2 
packaging==24.1 
regex==2024.9.11 
safetensors==0.4.5 
tokenizers==0.20.1 
transformers==4.45.2
MarkupSafe==3.0.1 
jinja2==3.1.4 
mpmath==1.3.0 
networkx==3.4.1 
# nvidia-cublas-cu12==12.1.3.1 
# nvidia-cuda-cupti-cu12==12.1.105 
# nvidia-cuda-nvrtc-cu12==12.1.105 
# nvidia-cuda-runtime-cu12==12.1.105 
# nvidia-cudnn-cu12==9.1.0.70 
# nvidia-cufft-cu12==11.0.2.54 
# nvidia-curand-cu12==10.3.2.106 
# nvidia-cusolver-cu12==11.4.5.107 
# nvidia-cusparse-cu12==12.1.0.106 
# nvidia-nccl-cu12==2.20.5 
# nvidia-nvjitlink-cu12==12.6.77 
# nvidia-nvtx-cu12==12.1.105 
pillow==10.4.0 
sympy==1.13.3 
torch==2.4.1 
torchaudio==2.4.1 
torchvision==0.19.1 
# triton==3.0.0
gunicorn==23.0.0